requirements.txt
langchain
langgraph
langchain-openai
requests
beautifulsoup4
prompt_toolkit
python-pptx
rich
